URL_Scanner/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
dnspython==2.5.0
whois==0.9.27
python-whois==0.9.27
requests==2.31.0
beautifulsoup4==4.12.3
tldextract==5.1.1
ipwhois==1.2.0
python-dateutil==2.8.2
geoip2==4.8.0
python-dotenv==1.0.1
